pyproject.toml
[project]
name = "jtg-spec-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.11.1",
    "celery>=5.6.3",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-login>=0.6.3",
    "flask-restful>=0.3.10",
    "gevent>=26.4.0",
    "json-repair>=0.59.4",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.13",
    "langgraph>=1.1.6",
    "py2neo>=2021.2.4",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
